Make Parser, begin generator

+++ src/tokesizer.hxx
@@ -109,5 +109,5 @@ class Tokenizer {
         }
 
         const std::string m_src;
-        int m_index = 0;
+        size_t m_index = 0;
 };
